Move FilterEventPredicate to model.event
