+++ src/main/java/seedu/address/logic/commands/FilterEventCommand.java
@@ -1,14 +1,12 @@
 package seedu.address.logic.commands;
 
 import java.util.List;
-import java.util.Set;
-import java.util.function.Predicate;
 
 import seedu.address.commons.core.index.Index;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
 import seedu.address.model.event.Event;
-import seedu.address.model.person.Person;
+import seedu.address.model.event.FilterEventPredicate;
 
 public class FilterEventCommand extends Command {
     public static final String COMMAND_WORD = "filter-event";
@@ -39,18 +37,5 @@ public class FilterEventCommand extends Command {
         FilterEventPredicate predicate = new FilterEventPredicate(event.getAttendeesList());
         model.updateFilteredPersonList(predicate);
         return new CommandResult(String.format(MESSAGE_FILTER_SUCCESS, event.getName()));
-    }
-
-    private static class FilterEventPredicate implements Predicate<Person> {
-        private final Set<Person> attendeeList;
-
-        public FilterEventPredicate(Set<Person> attendeeList) {
-            this.attendeeList = attendeeList;
-        }
-
-        @Override
-        public boolean test(Person person) {
-            return attendeeList.contains(person);
-        }
     }
 }
